Fixing compilation if HPX_WITH_THREAD_BACKTRACE_ON_SUSPENSION=On
